Add VAX/VMS CONOLD for 1989-MAR
The *.DECW$BOOKSHELF files in the [DECW$BOOK] directory are all that is needed to indetify all the manuals on this dics.

The subdirectory name should end DECW$BOOK but Linux doesn't like an unescaped dollar, so I've decided that the easiest thing to do is to drop the dollar.

Adjust the *.DECW$BOOKSHELF files to have a suitable filename extensionin the same way.