First pass at turning raw SPD text into YAML metadata
processing/spds-to-yaml.go: golang code to process one of more SPD text files and extract relevant metadata (SPD ID, title, DEC part number , date).

cd-contents/CDROMs.yaml: YAML that represents a number of DEC CDROM discs and allows spds-to-yaml to work out from the SPD text file subdirectory exactly which disc the SPD came from. This requires that the SPDs be organised in subdirectories with an expected name.

processing/months.yaml: SPDs list the date in the form "March 1989" and is transformed to "1989-03" in the metadata. The month name can be in a number of languages. Rather than hard-coding that in the golang code, it is stored as YAML and read in during execution of the program, making it easy to extend as other languages are found.