spds-to-yaml.go: Improve a message and try to fix a regexp
When indicating that too many SPDs have been found, say which SPD file this message refers to and include at least the first few lines of the potential title text.

WPLSWEDISH041_SPD.TXT on the Nov 1992 CONDIST pairs PRODUCT NAME with BESKRIVNING so try to allow for that. It still fails, but I've yet to chase down exactly why.