spds-to-yaml.go: Output rejected SPD files in YAML format
To make it easier to spot issues, a list of "SPDs with known issues" will be supplied.
To start with this can be the set of currently available SPDs that cannot currently be processed .
That amounts to just over 200 files at the moment.

Some of these files cannot be read at all because of their encoding (but perhaps they can be read in the future once that encoding has been determined).
In a few cases, the TXT file is currently indecipherable, but a PS file exists that can be read.

Other files can be read but simply do not have an SPD anywhere in them. For a few files the information is present but in a format that makes
it hard for the analysis code to find it.

This "known issues" list can be adjusted over time to fill in the blanks that can be gleaned by hand, allowing this program to produce a more
comprehensive list while still automatically analysing the vast majority of the SPD files.